+++ parsedmarc/splunk.py
@@ -1,10 +1,16 @@
+import logging
 from urllib.parse import urlparse
 import socket
 import json
+import urllib3
 
 import requests
 
 from parsedmarc import __version__, human_timestamp_to_timestamp
+
+urllib3.disable_warnings(urllib3.exceptions.InsecureRequestWarning)
+
+logger = logging.getLogger("parsedmarc")
 
 
 class SplunkError(RuntimeError):
@@ -54,6 +60,7 @@ class HECClient(object):
             to save in Splunk
 
         """
+        logger.debug("Saving aggregate reports to Splunk")
         if type(aggregate_reports) == dict:
             aggregate_reports = [aggregate_reports]
 
@@ -115,6 +122,7 @@ class HECClient(object):
             to save in Splunk
 
         """
+        logger.debug("Saving forensic reports to Splunk")
         if type(forensic_reports) == dict:
             forensic_reports = [forensic_reports]